requirements.txt
streamlit
flask
requests
pillow
